Fix Activities

+++ marlo-web/src/main/java/org/cgiar/ccafs/marlo/data/manager/impl/ActivityManagerImpl.java
@@ -22,12 +22,12 @@ import org.cgiar.ccafs.marlo.data.dao.PhaseDAO;
 import org.cgiar.ccafs.marlo.data.dao.ProjectDAO;
 import org.cgiar.ccafs.marlo.data.dao.ProjectPartnerPersonDAO;
 import org.cgiar.ccafs.marlo.data.manager.ActivityManager;
+import org.cgiar.ccafs.marlo.data.manager.DeliverableActivityManager;
 import org.cgiar.ccafs.marlo.data.model.Activity;
 import org.cgiar.ccafs.marlo.data.model.DeliverableActivity;
 import org.cgiar.ccafs.marlo.data.model.Phase;
 import org.cgiar.ccafs.marlo.data.model.ProjectPartnerPerson;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -43,23 +43,23 @@ public class ActivityManagerImpl implements ActivityManager {
 
 
   private ActivityDAO activityDAO;
-
   private PhaseDAO phaseDAO;
   private ProjectDAO projectDAO;
   private DeliverableActivityDAO deliverableActivityDAO;
-
+  private DeliverableActivityManager deliverableActivityManager;
   private ProjectPartnerPersonDAO projectPartnerPersonDAO;
-  // Managers
 
 
   @Inject
   public ActivityManagerImpl(ActivityDAO activityDAO, PhaseDAO phaseDAO, ProjectDAO projectDAO,
-    DeliverableActivityDAO deliverableActivityDAO, ProjectPartnerPersonDAO projectPartnerPersonDAO) {
+    DeliverableActivityDAO deliverableActivityDAO, ProjectPartnerPersonDAO projectPartnerPersonDAO,
+    DeliverableActivityManager deliverableActivityManager) {
     this.activityDAO = activityDAO;
     this.phaseDAO = phaseDAO;
     this.projectDAO = projectDAO;
     this.deliverableActivityDAO = deliverableActivityDAO;
     this.projectPartnerPersonDAO = projectPartnerPersonDAO;
+    this.deliverableActivityManager = deliverableActivityManager;
   }
 
   /**
@@ -87,11 +87,10 @@ public class ActivityManagerImpl implements ActivityManager {
    * 
    * @param deliverableActivityAdd deliverableActivity to clone
    * @param deliverableActivity base deliverableActivity
-   * @param activity the activity base
    * @param activityAdd the new activity base
    */
   public void cloneDeliverableActivity(DeliverableActivity deliverableActivityAdd,
-    DeliverableActivity deliverableActivity, Activity activity, Activity activityAdd, Phase phase) {
+    DeliverableActivity deliverableActivity, Activity activityAdd, Phase phase) {
     deliverableActivityAdd.setActivity(activityAdd);
     deliverableActivityAdd.setDeliverable(deliverableActivity.getDeliverable());
     deliverableActivityAdd.setPhase(phase);
@@ -112,7 +111,7 @@ public class ActivityManagerImpl implements ActivityManager {
     if (activity.getDeliverables() != null) {
       for (DeliverableActivity deliverableActivity : activity.getDeliverables()) {
         DeliverableActivity deliverableActivityAdd = new DeliverableActivity();
-        this.cloneDeliverableActivity(deliverableActivityAdd, deliverableActivity, activity, activityAdd, phase);
+        this.cloneDeliverableActivity(deliverableActivityAdd, deliverableActivity, activityAdd, phase);
         deliverableActivityDAO.save(deliverableActivityAdd);
       }
     }
@@ -125,9 +124,10 @@ public class ActivityManagerImpl implements ActivityManager {
     List<Activity> activities =
       phase.getProjectActivites().stream().filter(c -> c.isActive() && c.getProject().getId().longValue() == projecID
         && activity.getComposeID().equals(c.getComposeID())).collect(Collectors.toList());
-    for (Activity activityDB : activities) {
-
-      activityDAO.deleteActivity(activityDB.getId());
+    if (activities != null) {
+      for (Activity activityDB : activities) {
+        activityDAO.deleteActivity(activityDB.getId());
+      }
     }
 
     if (phase.getNext() != null) {
@@ -141,10 +141,19 @@ public class ActivityManagerImpl implements ActivityManager {
 
     Activity activity = this.getActivityById(activityId);
     activityDAO.deleteActivity(activityId);
+
     Phase currentPhase = phaseDAO.find(activity.getPhase().getId());
     if (currentPhase.getDescription().equals(APConstants.PLANNING)) {
       if (activity.getPhase().getNext() != null) {
         this.deletActivityPhase(activity.getPhase().getNext(), activity.getProject().getId(), activity);
+      }
+    }
+    if (currentPhase.getDescription().equals(APConstants.REPORTING)) {
+      if (currentPhase.getNext() != null && currentPhase.getNext().getNext() != null) {
+        Phase upkeepPhase = currentPhase.getNext().getNext();
+        if (upkeepPhase != null) {
+          this.deletActivityPhase(upkeepPhase, activity.getProject().getId(), activity);
+        }
       }
     }
   }
@@ -191,21 +200,42 @@ public class ActivityManagerImpl implements ActivityManager {
 
     Activity resultActivity = activityDAO.save(activity);
     Phase currentPhase = phaseDAO.find(activity.getPhase().getId());
+
+    this.saveCurrentPhaseDeliverables(resultActivity, activity.getDeliverables(), currentPhase);
+
     if (currentPhase.getDescription().equals(APConstants.PLANNING)) {
       if (activity.getPhase().getNext() != null) {
         this.saveActvityPhase(activity.getPhase().getNext(), activity.getProject().getId(), activity);
       }
     }
+
+    if (currentPhase.getDescription().equals(APConstants.REPORTING)) {
+      if (currentPhase.getNext() != null && currentPhase.getNext().getNext() != null) {
+        Phase upkeepPhase = currentPhase.getNext().getNext();
+        if (upkeepPhase != null) {
+          this.saveActvityPhase(upkeepPhase, activity.getProject().getId(), activity);
+        }
+      }
+    }
     return resultActivity;
   }
 
+  /**
+   * Save activities for next phase
+   * 
+   * @param next
+   * @param projecID
+   * @param activity
+   */
   public void saveActvityPhase(Phase next, long projecID, Activity activity) {
     Phase phase = phaseDAO.find(next.getId());
 
     List<Activity> activities =
       phase.getProjectActivites().stream().filter(c -> c.isActive() && c.getProject().getId().longValue() == projecID
         && c.getComposeID().equals(activity.getComposeID())).collect(Collectors.toList());
-    if (activities.isEmpty()) {
+
+    // Create new Activity
+    if (activities == null || activities.isEmpty()) {
       Activity activityAdd = new Activity();
       this.cloneActivity(activityAdd, activity, phase);
       activityDAO.save(activityAdd);
@@ -214,48 +244,71 @@ public class ActivityManagerImpl implements ActivityManager {
         activityAdd.setComposeID(activity.getComposeID());
         activityDAO.save(activityAdd);
       }
-      if (activity.getDeliverables() != null) {
-        for (DeliverableActivity deliverableActivity : activity.getDeliverables()) {
-          DeliverableActivity deliverableActivityAdd = new DeliverableActivity();
-          this.cloneDeliverableActivity(deliverableActivityAdd, deliverableActivity, activity, activityAdd, phase);
-          deliverableActivityDAO.save(deliverableActivityAdd);
-        }
-      }
+      this.saveCurrentPhaseDeliverables(activityAdd, activity.getDeliverables(), phase);
     } else {
+      // Update activity
+
       Activity activityAdd = activities.get(0);
       this.cloneActivity(activityAdd, activity, phase);
       activityDAO.save(activityAdd);
-      if (activity.getDeliverables() == null) {
-        activity.setDeliverables(new ArrayList<DeliverableActivity>());
-      }
-      for (DeliverableActivity deliverableActivity : activity.getDeliverables()) {
-        if (activityAdd.getDeliverableActivities().stream()
-          .filter(c -> c.isActive() && c.getDeliverable().equals(deliverableActivity.getDeliverable()))
-          .collect(Collectors.toList()).isEmpty()) {
-          DeliverableActivity deliverableActivityAdd = new DeliverableActivity();
-          this.cloneDeliverableActivity(deliverableActivityAdd, deliverableActivity, activity, activityAdd, phase);
-          deliverableActivityDAO.save(deliverableActivityAdd);
-        }
-      }
-      for (DeliverableActivity deliverableActivity : activityAdd.getDeliverableActivities().stream()
-        .filter(c -> c.isActive()).collect(Collectors.toList())) {
-        if (activity.getDeliverables().stream()
-          .filter(c -> c.getDeliverable().equals(deliverableActivity.getDeliverable())).collect(Collectors.toList())
-          .isEmpty()) {
-          deliverableActivityDAO.deleteDeliverableActivity(deliverableActivity.getId());
-        }
-
-      }
-
-
+      this.saveCurrentPhaseDeliverables(activityAdd, activity.getDeliverables(), phase);
     }
 
-    if (phase.getNext() != null)
-
-    {
+    if (phase.getNext() != null) {
       this.saveActvityPhase(phase.getNext(), projecID, activity);
     }
 
+
+  }
+
+  /**
+   * Save/Delete activityDeliverable of the current phase
+   */
+  private void saveCurrentPhaseDeliverables(Activity activityUI, List<DeliverableActivity> deliverableActivitiesUI,
+    Phase currentPhase) {
+    List<DeliverableActivity> deliverableActivitiesDB =
+      activityUI.getDeliverableActivities().stream().filter(da -> da.isActive()).collect(Collectors.toList());
+    if (deliverableActivitiesUI != null) {
+
+      // Delete deliverableActivities distinct from DB
+      for (DeliverableActivity deliverableActivity : deliverableActivitiesDB) {
+        if (!deliverableActivitiesUI.contains(deliverableActivity)) {
+          deliverableActivityManager.deleteDeliverableActivity(deliverableActivity.getId());
+        }
+      }
+
+      // Add deliverableActivity if not exist
+      for (DeliverableActivity deliverableActivity : deliverableActivitiesUI) {
+
+        if (deliverableActivity.getId() == null || deliverableActivity.getId() == -1) {
+          // New DeliverableActivity
+          DeliverableActivity deliverableActivityNew = new DeliverableActivity();
+          this.cloneDeliverableActivity(deliverableActivityNew, deliverableActivity, activityUI, currentPhase);
+          deliverableActivityManager.saveDeliverableActivity(deliverableActivityNew);
+        } else {
+          // Check if already exists in DB, then save
+          List<DeliverableActivity> deliverableActivities = currentPhase.getDeliverableActivities().stream()
+            .filter(da -> da.isActive() && da.getActivity() != null
+              && da.getActivity().getId().longValue() == activityUI.getId().longValue()
+              && da.getDeliverable().equals(deliverableActivity.getDeliverable()))
+            .collect(Collectors.toList());
+          if (deliverableActivities == null || deliverableActivities.isEmpty()) {
+            DeliverableActivity deliverableActivityAdd = new DeliverableActivity();
+            this.cloneDeliverableActivity(deliverableActivityAdd, deliverableActivity, activityUI, currentPhase);
+            deliverableActivityDAO.save(deliverableActivityAdd);
+          }
+        }
+
+      }
+    } else {
+      // delete all from db
+      if (deliverableActivitiesDB != null && !deliverableActivitiesDB.isEmpty()) {
+        for (DeliverableActivity deliverableActivity : deliverableActivitiesDB) {
+          deliverableActivityManager.deleteDeliverableActivity(deliverableActivity.getId());
+        }
+      }
+
+    }
 
   }
 }
